back-end/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
openai==0.27.0
python-dotenv==0.21.0
python-multipart==0.0.5
pdfplumber==0.6.0
